requirements.txt
appdirs==1.4.3
appnope==0.1.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2015.11.20.1
cycler==0.10.0
Cython==0.23.4
decorator==4.0.11
future==0.16.0
gnureadline==6.3.3
imageio==2.1.2
ipykernel==4.2.2
ipython==4.0.1
ipython-genutils==0.1.0
ipywidgets==4.1.1
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
MarkupSafe==0.23
matplotlib==2.0.0
mistune==0.7.1
moviepy==0.2.3.2
nbconvert==4.1.0
nbformat==4.0.1
notebook==4.0.6
numpy==1.12.1
opencv-python==3.2.0.7
packaging==16.8
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
ptyprocess==0.5
Pygments==2.0.2
pyparsing==2.0.7
python-dateutil==2.4.2
pytube==6.2.2
pytz==2015.7
pyzmq==15.1.0
qtconsole==4.1.1
scipy==0.16.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
site==0.0.1
six==1.10.0
terminado==0.5
tornado==4.3
tqdm==4.11.2
traitlets==4.0.0
pandas==0.19.2
tensorflow-gpu
keras==2.0.4
h5py==2.5.0
